Put logic in BrevService instead of route
