+++ web/src/main/kotlin/no/nav/su/se/bakover/web/routes/behandling/BehandlingRoutes.kt
@@ -15,7 +15,6 @@ import io.ktor.routing.Route
 import io.ktor.routing.get
 import io.ktor.routing.patch
 import io.ktor.routing.post
-import no.nav.su.se.bakover.client.dokarkiv.DokArkiv
 import no.nav.su.se.bakover.client.person.PersonOppslag
 import no.nav.su.se.bakover.database.ObjectRepo
 import no.nav.su.se.bakover.domain.AktørId
@@ -46,7 +45,6 @@ internal fun Route.behandlingRoutes(
     personOppslag: PersonOppslag,
     oppgaveClient: OppgaveClient,
     utbetalingPublisher: UtbetalingPublisher,
-    dokArkiv: DokArkiv
 ) {
     val log = LoggerFactory.getLogger(this::class.java)
 
@@ -133,27 +131,12 @@ internal fun Route.behandlingRoutes(
         // TODO authorize attestant
         call.withBehandling(repo) { behandling ->
             call.audit("Attesterer behandling med id: ${behandling.id}")
-
             val sak = repo.hentSak(behandling.sakId) ?: throw RuntimeException("Sak id finnes ikke")
-            val person = personOppslag.person(sak.fnr).getOrElse { throw RuntimeException("Finner ikke person") }
-            val pdf = brevService.lagUtkastTilBrev(behandling.toDto()).getOrElse {
-                throw RuntimeException("Kunne ikke lage utkast av brev")
-            }
 
-            dokArkiv.opprettJournalpost(
-                dokumentInnhold = BrevService.lagVedtakInnhold(person = person, behandlingDto = behandling.toDto()),
-                person = person,
-                pdf = pdf,
-                sakId = sak.id.toString()
-            ).fold(
-                ifLeft = {
-                    log.error("Kunne ikke opprette journalpost for attestering")
-                    call.svar(InternalServerError.message("Kunne ikke opprette journalpost for attestering"))
-                },
-                ifRight = { journalPostId -> brevService.sendBrev(journalPostId) }
+            brevService.oprettJournalpostOgSendBrev(sak, behandling.toDto()).fold(
+                ifLeft = { call.svar(InternalServerError.message("Feilet ved attestering")) },
+                ifRight = { call.svar(OK.jsonBody(behandling.attester(attestant = Attestant(id = call.lesBehandlerId())))) }
             )
-
-            call.svar(OK.jsonBody(behandling.attester(attestant = Attestant(id = call.lesBehandlerId()))))
         }
     }
